Implemented the new Event objects. Not yet operational
