+++ src/com/merendabot/commands/commands/ClassesCommand.java
@@ -2,10 +2,10 @@ package com.merendabot.commands.commands;
 
 import com.merendabot.commands.CommandCategory;
 import com.merendabot.commands.CommandClass;
+import com.merendabot.university.events.Class;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import com.merendabot.university.Merenda;
-import com.merendabot.university.events.Event;
 import com.merendabot.university.events.EventType;
 import com.merendabot.university.subjects.Subject;
 
@@ -33,16 +33,15 @@ public class ClassesCommand extends CommandClass {
                 String fieldTitle = dayOfWeek.getDisplayName(TextStyle.FULL_STANDALONE, Locale.forLanguageTag("PT"));
                 StringBuilder fieldValue = new StringBuilder();
 
-                for (Event e : Event.getEventsByWeekday(dayOfWeek, EventType.CLASS)) {
-                    Subject subject = Subject.getSubjectById(e.getSubjectId());
+                for (Class c : Class.getClassesByWeekDay(dayOfWeek)) {
                     fieldValue.append(
                             String.format(
                                     "**%s %s** (%s - %s) - [Link](%s)%n",
-                                    e.getName(),
-                                    subject.getShortName(),
-                                    e.getStartTime().format(DateTimeFormatter.ofPattern("HH:mm")),
-                                    e.getEndTime().format(DateTimeFormatter.ofPattern("HH:mm")),
-                                    e.getLink()));
+                                    c.getName(),
+                                    c.getSubject().getShortName(),
+                                    c.getStartTime().format(DateTimeFormatter.ofPattern("HH:mm")),
+                                    c.getEndTime().format(DateTimeFormatter.ofPattern("HH:mm")),
+                                    c.getLink()));
                 }
                 if (fieldValue.length() > 0) // Filter days that have no classes
                     eb.addField(fieldTitle, fieldValue.toString(), false);
